Move main navigation widgets to constructor
